requirements.txt
datasets
huggingface_hub
tiktoken
openai
anthropic
google-genai
